Centralizing configs so they are more streamlined for the client
